docs(form): Improve form stories
